Move persistence types from Integration -> ConnectedMode (#4327)
* Move persistence types from Integration -> ConnectedMode

TODO - fix namespaces

Part of #4320
