+++ src/ConnectedMode.UnitTests/Migration/ObsoleteConfigurationProviderTests.cs
