+++ src/ConnectedMode.UnitTests/Migration/ObsoleteConnectedModeSolutionBindingPathProviderTests.cs
