+++ src/ConnectedMode.UnitTests/Persistence/BoundSonarQubeProjectExtensionsTests.cs
