+++ src/ConnectedMode.UnitTests/Persistence/ConfigurationPersisterTests.cs
