+++ src/ConnectedMode.UnitTests/Persistence/SolutionBindingCredentialsLoaderTests.cs
@@ -19,10 +19,7 @@
  */
 
 using System;
-using FluentAssertions;
 using Microsoft.Alm.Authentication;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Moq;
 using SonarLint.VisualStudio.ConnectedMode.Persistence;
 using SonarLint.VisualStudio.Core.Binding;
 using SonarLint.VisualStudio.Integration.Persistence;
