+++ src/ConnectedMode.UnitTests/Persistence/SolutionBindingDataReaderTests.cs
