+++ src/ConnectedMode.UnitTests/Persistence/SolutionBindingDataWriterTests.cs
