+++ src/ConnectedMode.UnitTests/Persistence/SolutionBindingFileLoaderTests.cs
