+++ src/ConnectedMode/Binding/CredentialStore.cs
