+++ src/ConnectedMode/Binding/ICredentialStoreService.cs
