+++ src/ConnectedMode/Binding/UnintrusiveBindingPathProvider.cs
@@ -56,7 +56,7 @@ namespace SonarLint.VisualStudio.Integration.UnintrusiveBinding
             SLVSRootBindingFolder = Path.Combine(environmentVariables.GetSLVSAppDataRootPath(), "Bindings");
 
             IVsSolution slnService = null;
-            threadHandling.RunOnUIThreadSync(() => slnService = serviceProvider.GetService<SVsSolution, IVsSolution>());
+            threadHandling.RunOnUIThreadSync(() => slnService = serviceProvider.GetService(typeof(SVsSolution)) as IVsSolution);
             solution = slnService;
         }
