+++ src/ConnectedMode/ConnectedMode.csproj
@@ -53,6 +53,11 @@
       <AutoGen>True</AutoGen>
       <DependentUpon>BindingStrings.resx</DependentUpon>
     </Compile>
+    <Compile Update="Persistence\PersistenceStrings.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>PersistenceStrings.resx</DependentUpon>
+    </Compile>
     <Compile Update="Resources.Designer.cs">
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
@@ -64,6 +69,10 @@
     <EmbeddedResource Update="Binding\BindingStrings.resx">
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>BindingStrings.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+    <EmbeddedResource Update="Persistence\PersistenceStrings.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>PersistenceStrings.Designer.cs</LastGenOutput>
     </EmbeddedResource>
     <EmbeddedResource Update="Resources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
