+++ src/ConnectedMode/Migration/ObsoleteConfigurationProvider.cs
