+++ src/ConnectedMode/Persistence/ISolutionBindingCredentialsLoader.cs
