+++ src/ConnectedMode/Persistence/ISolutionBindingDataReader.cs
