+++ src/ConnectedMode/Persistence/ISolutionBindingDataWriter.cs
