+++ src/ConnectedMode/Persistence/ISolutionBindingFileLoader.cs
