+++ src/ConnectedMode/Persistence/ISolutionBindingPathProvider.cs
