+++ src/ConnectedMode/Persistence/SolutionBindingCredentialsLoader.cs
